# specs/001-job-search-mvp/contracts/openapi.yml
openapi: 3.1.0
info:
  title: Job Enhancer API
  description: Job Search Assistant MVP — FastAPI backend
  version: 1.0.0
paths:
...
  /v1/saved-searches/:
    get:
      tags:
      - SavedSearches
      summary: List Saved Searches
      operationId: list_saved_searches_v1_saved_searches__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SavedSearchSchema'
                type: array
                title: Response List Saved Searches V1 Saved Searches  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - SavedSearches
      summary: Create Saved Search
      operationId: create_saved_search_v1_saved_searches__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/saved-searches/matches:
    get:
      tags:
      - SavedSearches
      summary: Get New Matches
      description: Jobs found since each saved search was last viewed.
      operationId: get_new_matches_v1_saved_searches_matches_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewMatchesResponse'
      security:
      - HTTPBearer: []
  /v1/saved-searches/mark-seen:
    post:
      tags:
      - SavedSearches
      summary: Mark Matches Seen
      operationId: mark_matches_seen_v1_saved_searches_mark_seen_post
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
  /v1/saved-searches/{search_id}:
    delete:
      tags:
      - SavedSearches
      summary: Delete Saved Search
      operationId: delete_saved_search_v1_saved_searches__search_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: search_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Search Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/users/me:
    get:
      tags:
      - Users
      summary: Get Profile
      operationId: get_profile_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - Users
      summary: Delete Account
      description: 'Delete the current user account (FR-020).


        The account is deactivated immediately; all owned rows and stored

        files are permanently purged after a 30-day grace period by the

        daily `purge_deleted_users` job (cascade via FK).'
      operationId: delete_account_v1_users_me_delete
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Users
      summary: Update Profile
      operationId: update_profile_v1_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/users/me/export:
    get:
      tags:
      - Users
      summary: Export Data
      description: Serialize all user-owned data as a JSON download.
      operationId: export_data_v1_users_me_export_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobListingSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        external_id:
          type: string
          title: External Id
        source:
          type: string
          title: Source
        title:
          type: string
          title: Title
        company:
          type: string
          title: Company
        location:
          type: string
          title: Location
        is_remote:
          type: boolean
          title: Is Remote
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        salary_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Min
        salary_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Max
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
        job_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Type
        apply_url:
          type: string
          title: Apply Url
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        is_expired:
          type: boolean
          title: Is Expired
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - external_id
      - source
      - title
      - company
      - location
      - is_remote
      - description
      - salary_min
      - salary_max
      - currency
      - job_type
      - apply_url
      - posted_at
      - is_expired
      - created_at
      title: JobListingSchema
...
    NewMatchesResponse:
      properties:
        matches:
          items:
            $ref: '#/components/schemas/SearchMatches'
          type: array
          title: Matches
        total_new:
          type: integer
          title: Total New
      type: object
      required:
      - matches
      - total_new
      title: NewMatchesResponse
...
    SavedSearchCreate:
      properties:
        q:
          type: string
          maxLength: 255
          minLength: 1
          title: Q
        location:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Location
        remote_only:
          type: boolean
          title: Remote Only
          default: false
        salary_min:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Salary Min
        salary_max:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Salary Max
        experience:
          anyOf:
          - type: string
            pattern: ^(entry|mid|senior)$
          - type: 'null'
          title: Experience
        job_type:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Job Type
        name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Name
      type: object
      required:
      - q
      title: SavedSearchCreate
    SavedSearchSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        q:
          type: string
          title: Q
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        remote_only:
          type: boolean
          title: Remote Only
        salary_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Min
        salary_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Max
        experience:
          anyOf:
          - type: string
          - type: 'null'
          title: Experience
        job_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Type
        is_active:
          type: boolean
          title: Is Active
        last_run_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Run At
        last_viewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Viewed At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - q
      - location
      - remote_only
      - salary_min
      - salary_max
      - experience
      - job_type
      - is_active
      - last_run_at
      - last_viewed_at
      - created_at
      title: SavedSearchSchema
    SearchMatches:
      properties:
        search:
          $ref: '#/components/schemas/SavedSearchSchema'
        new_jobs:
          items:
            $ref: '#/components/schemas/JobListingSchema'
          type: array
          title: New Jobs
        new_count:
          type: integer
          title: New Count
      type: object
      required:
      - search
      - new_jobs
      - new_count
      title: SearchMatches
...
    UserProfile:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        role:
          type: string
          title: Role
        follow_up_days:
          type: integer
          title: Follow Up Days
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - role
      - follow_up_days
      - created_at
      title: UserProfile
    UserUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        follow_up_days:
          anyOf:
          - type: integer
            maximum: 90.0
            minimum: 1.0
          - type: 'null'
          title: Follow Up Days
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local development
- url: https://job-enhancer-api.onrender.com
  description: Production (Render)